docs(initial): First draft of DS Technology Tree
This is the very first version with a couple of different ideas that I've
taken from napkins. There's still a lot to be desired here, but it's enough
of a starting point that I can start to share it with other people and accept
feedback.

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>